Nvidia-Scalling_Down-DeployReady/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
binaryornot==0.4.4
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
hpack==4.0.0
keras==3.5.0
libclang==18.1.1
Markdown==3.7
matplotlib==3.9.2
ml-dtypes==0.3.2
munkres==1.1.4
namex==0.0.8
opt_einsum==3.4.0
optree==0.12.1
protobuf==4.25.5
pyarrow==17.0.0
sentencepiece==0.2.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==2.4.0
tokenizers==0.20.0
torchaudio==2.4.0
transformers==4.45.1
Werkzeug==3.0.4
wrapt==1.16.0
zstandard==0.23.0

numpy~=1.26.4
yaml~=0.2.5
pyyaml~=6.0.2
tqdm~=4.66.5
scikit-learn~=1.5.1
datasets~=3.0.1
pandas~=2.2.3